Dynamic help text and cleanup
I added a secondary observer pattern to the Help command to dynamically add the command text from the commendable objects.  I also removed the unneeded List and Sort commands for now as they are not fully implemented.  And did some general cleanup in the MyGame class and updated the command objects.  I also added the index integer to the inventory list display command.